refactor: delete legacy threshold/protocols/corona Feldman-DKG impl (2292 LOC superseded by luxfi/corona's corrected R-LWE construction)
